Chaning unsigned long to uint64_t
Double-checked the 'long long' one and it's good now

+++ src/comp_management.h
@@ -115,7 +115,7 @@ public:
   {
     std::mt19937_64 eng(config_.randomseed); //Use the 64-bit Mersenne Twister 19937 generator
                                //and seed it with entropy.
-    std::uniform_int_distribution<unsigned long long> distr;
+    std::uniform_int_distribution<uint64_t> distr;
     assert(seedforCLHASH.empty());
     seedforCLHASH.resize(config_.hashrange);
     for (unsigned i = 0; i < config_.hashrange; i++) {
